+++ src/modules/esm.mjs
@@ -1,0 +1,37 @@
+import path from "path";
+import fs from "fs/promises";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import "./files/c.cjs";
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = await fs.readFile("src/modules/files/a.json", "utf8");
+} else {
+  unknownObject = await fs.readFile("src/modules/files/b.json", "utf8");
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${import.meta.url}`);
+console.log(`Path to current directory is ${new URL(".", import.meta.url)}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
